requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
PyJWT==2.8.0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0.post1
watchfiles==1.1.1
websockets==16.0
